Refactored test classes simplification
