+++ src/main/java/com/github/dsmiles/awslocal/message/Message.java
@@ -1,4 +1,4 @@
-package com.github.dsmiles.awslocal;
+package com.github.dsmiles.awslocal.message;
 
 import java.util.UUID;
