+++ src/main/java/com/github/dsmiles/awslocal/message/MessageController.java
@@ -1,5 +1,7 @@
-package com.github.dsmiles.awslocal;
+package com.github.dsmiles.awslocal.message;
 
+import com.github.dsmiles.awslocal.properties.ApplicationProperties;
+import com.github.dsmiles.awslocal.storage.StorageService;
 import org.springframework.web.bind.annotation.*;
 
 import java.io.IOException;
