+++ src/main/java/com/github/dsmiles/awslocal/properties/ApplicationProperties.java
@@ -1,4 +1,4 @@
-package com.github.dsmiles.awslocal;
+package com.github.dsmiles.awslocal.properties;
 
 import org.springframework.boot.context.properties.ConfigurationProperties;
